pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["NNbaur <studentsoftpu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.6"
dj-database-url = "^0.5.0"
python-dotenv = "^0.20.0"
gunicorn = "^20.1.0"
django-bootstrap4 = "^22.1"
psycopg2= "^2.9.3"
whitenoise = "6.2.0"
django-filter = "^22.1"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
